Create a SDK for Flickr
I've extracted this SDK/client for Flickr from the main project as it
doesn't belong to the core of the application, this means, that if
instead of Flickr we used another photo gallery service, the app
should behave in the same way.

This only adds support to search for images based on terms, I've tried
to fetch Flickr landing page photos, but either isn't available or I
couldn't really figure it out, which is the service method that returns
the explorer version that they have on their own homepage.

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>